% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_correlation}
\alias{calculate_correlation}
\title{Calculate the sample correlation coefficient of two vectors}
\usage{
calculate_correlation(A, B)
}
\arguments{
\item{A}{Vector of dimension n}

\item{B}{Vector of dimension n}
}
\value{
Sample correlation coefficient of A and B
}
\description{
Calculate the sample correlation coefficient of two vectors
}
\examples{
calculate_correlation(c(1,2,3),c(3,2,1))
}
