% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_correlation_threshold_matrix}
\alias{calculate_correlation_threshold_matrix}
\title{Calculate deletion correlation thresholds for each pair of covariates
according to Section 8.3.2 of the thesis}
\usage{
calculate_correlation_threshold_matrix(Z, data_size)
}
\arguments{
\item{Z}{A matrix containing the selected covariates by the selection
procedure of Section 8.3.1 of dimension nxs}

\item{data_size}{The sample size n}
}
\value{
The matrix containing the thresholds for each pair of covariates
}
\description{
Calculate deletion correlation thresholds for each pair of covariates
according to Section 8.3.2 of the thesis
}
