% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_correlation_thresholds}
\alias{calculate_correlation_thresholds}
\title{Calculate correlation thresholds for each covariate with the outcome
according to Section 8.3.1 of the thesis}
\usage{
calculate_correlation_thresholds(Z, Y, data_size)
}
\arguments{
\item{Z}{The covariate matrix of dimension nxp}

\item{Y}{The outcome vector of dimension n}

\item{data_size}{The sample size n}
}
\value{
The threshold vector containing a threshold for each of the
        covariates
}
\description{
Calculate correlation thresholds for each covariate with the outcome
according to Section 8.3.1 of the thesis
}
