% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cross_interactions}
\alias{cross_interactions}
\title{Compute cross interaction terms of two sets of covariates}
\usage{
cross_interactions(Z1, Z2, ident = "")
}
\arguments{
\item{Z1}{The covariate matrix 1 of dimension n x p1}

\item{Z2}{The covariate matrix 2 of dimension n x p2}

\item{ident}{String for naming the resulting columns}
}
\value{
All cross interaction terms structured as a matrix
}
\description{
Compute cross interaction terms of two sets of covariates
}
