% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fourier_basis}
\alias{fourier_basis}
\title{Calculate trigonometric transformations of the covariates}
\usage{
fourier_basis(Z, order)
}
\arguments{
\item{Z}{The covariate matrix of dimension nxp}

\item{order}{The order up to which the transformations should be calculated}
}
\value{
The trigonometric transformations structured as a matrix
}
\description{
Calculate trigonometric transformations of the covariates
}
