% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{interaction_terms}
\alias{interaction_terms}
\title{Compute interaction terms of covariates}
\usage{
interaction_terms(Z)
}
\arguments{
\item{Z}{The covariate matrix of dimension nxp}
}
\value{
All interaction terms structured as matrix
}
\description{
Compute interaction terms of covariates
}
