% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDD_functions.R
\name{perform_rdd_on_data}
\alias{perform_rdd_on_data}
\title{Execute a RD analysis on a given data set including covariates chosen by
different selection procedures (used to generate results of Section 9.4)}
\usage{
perform_rdd_on_data(X, Y, Z, rdd_library = "honest", estimator_type = "1")
}
\arguments{
\item{X}{The running variable array of dimension n}

\item{Y}{The outcome array of dimension n}

\item{Z}{The covariate matrix of dimension nxp}

\item{rdd_library}{The R package to use for the RD analysis. Possible values
are "honest" for the package RDHonest and "robust" for
the package RDRobust.}

\item{estimator_type}{This parameter is just relevant when using RDRobust
(otherwise it can be ignored). It indicates the
estimator type used in the RD analysis.
Possible values are:
1 - for conventional estimator,
2 - for bias-corrected estimator,
3 - for robust estimator}
}
\value{
The results of the RD analysis (estimation, standard error,
        confidence intervals) as well as the results on the selection of
        covariates
}
\description{
Execute a RD analysis on a given data set including covariates chosen by
different selection procedures (used to generate results of Section 9.4)
}
