% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDD_functions.R
\name{perform_rdd_redundant_covariates}
\alias{perform_rdd_redundant_covariates}
\title{Execute a RD analysis on generated data including covariates chosen from a
set of redundant covariates by different selection procedures (used to
generate results of Section 9.3)}
\usage{
perform_rdd_redundant_covariates(
  run,
  sample_size,
  rdd_library = "honest",
  estimator_type = 1
)
}
\arguments{
\item{run}{The number of execution when executed multiple times in parallel}

\item{sample_size}{The sample size}

\item{rdd_library}{The R package to use for the RD analysis. Possible values
are "honest" for the package RDHonest and "robust" for
the package RDRobust.}

\item{estimator_type}{This parameter is just relevant when using RDRobust
(otherwise it can be ignored). It indicates the
estimator type used in the RD analysis.
Possible values are:
1 - for conventional estimator,
2 - for bias-corrected estimator,
3 - for robust estimator}
}
\value{
The results of the RD analysis (estimation, bias, standard deviation,
        standard error, confidence intervals, coverage) and
        the results on the selection of covariates
}
\description{
Execute a RD analysis on generated data including covariates chosen from a
set of redundant covariates by different selection procedures (used to
generate results of Section 9.3)
}
