% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_covs_calculated_threshold}
\alias{remove_covs_calculated_threshold}
\title{Remove selected covariates according to the procedure described in Section
8.3.2 and 8.3.3 (simple and advanced deletion, respectively)}
\usage{
remove_covs_calculated_threshold(Z, Y, data_size, simple_deletion = TRUE)
}
\arguments{
\item{Z}{A matrix containing the selected covariates by the Selection
procedure of Section 8.3.1 of dimension nxs}

\item{Y}{The outcome vector of dimension of n}

\item{data_size}{The sample size n}

\item{simple_deletion}{When set to TRUE, simple deletion is applied (Section
8.3.2), otherwise advanced deletion (Section 8.3.3)}
}
\value{
The remaining covariates after deletion
}
\description{
Remove selected covariates according to the procedure described in Section
8.3.2 and 8.3.3 (simple and advanced deletion, respectively)
}
