% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{triangular}
\alias{triangular}
\title{Compute the triangular kernel}
\usage{
triangular(x)
}
\arguments{
\item{x}{The argument passed to the triangular kernel function.}
}
\value{
The value of the triangular kernel function.
}
\description{
Compute the triangular kernel
}
