% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compare_correlation}
\alias{compare_correlation}
\title{Calculate the sample correlation coefficient of each covariate with the
outcome and check whether it is greater or equal a threshold}
\usage{
compare_correlation(Z, Y, threshold)
}
\arguments{
\item{Z}{The covariates matrix of dimension nxp}

\item{Y}{The outcome vector of dimension n}

\item{threshold}{The threshold vector of dimension n}
}
\value{
Returns indices of covariates which have a sample correlation to the
        outcome greater or equal the according threshold
}
\description{
Calculate the sample correlation coefficient of each covariate with the
outcome and check whether it is greater or equal a threshold
}
